Move severities to config, tidy up even more
